Remove creation date from genrated PDFs
The creation date included in the generated PDFs causes `npm run build`
to update all PDFs with a new creation date, causing a lot of
unnecessary changes. This hacks the CreateDate value to be an empty
string, preventing regeneration of the PDFs from causing changes to all
PDFs.

This can be tested by running `npm run build` on the master branch
and observing that all PDFs are changed. Then running `npm run build` on
this branch, commiting the result, running `npm run build` again and
observering no changes.